GENERATORS: Functions are no longer inlined
Functions declared inline should be an exception. Hence they
require special attention (for writing/maintaining/linking),
which basically means they are a burden on the developer.
Inline functions are no longer used in this project - we now
rely on smart compilers and global optimization. If it does not
work, we will change it...but in the optimization stage.

Signed-off-by: Gama <misept.dieseneuf@gmail.com>