GENERATORS: Changed the generators base class
Generators are now considered as "dumb" - no parameters management at all.
They apply the parameters as soon as they have it; it means that all
this management is done in the Vco wrapper.

Also all generators are supposed to have an instantaneous phase that can be
set/get at any moment.

The base class was renamed from "Interface" since it might no longer be an
interface if the need eventually arises.

Signed-off-by: Gama <misept.dieseneuf@gmail.com>