TESTS: Changed performance tests
The previous ones were relying on ComputeMean function (used to prevent
a too smart compielr to optimize an empty loop away).
It meant that the test also timed ComputeMean function implementation,
which makes this test not consistent accross implementations.

This way it should be better, only having a single comparison to do,
which is OK to vectorize.

Signed-off-by: Gama <misept.dieseneuf@gmail.com>